Give the HLF_* family one home
`HLF_*` — the index into `highlight_attr[]` naming each builtin UI
highlight group — was re-declared as a per-file `pub const` block in 44
modules, 1,714 declarations of 77 names. c2rust emitted them as
`c_uint` (the C enum's type), while essentially every consumer wrote
`HLF_X as c_int`, so the family also carried ~250 casts.

The canonical home is `highlight_group/hlf.rs`, re-exported from
`highlight_group` so the family reads as one item of that module's API:
highlight_group owns `hlf_names`, the table the indices resolve against,
and is the only module that iterates the whole range. Doc comments come
from `v0.12.4:src/nvim/highlight_defs.h`, paired with the group name each
index resolves to — `HLF_8`/`HLF_AT`/`HLF_0` say nothing on their own.

Typed to the consumer (`c_int`), which is also what `hlf_T` had to
become: 71 sites assign an `HLF_*` straight into an `hlf_T` local or
field, and every one of them then casts to `c_int` to use it.

consolidate.py did the deletion and the imports (with `--eval`, so the
values were compared numerically, not textually); the four `const HLF_x:
c_int = super::HLF_x as c_int` aliases in search/'s children and one
`HLF_T` hiding in undo.rs's indented `mod header` were removed by hand,
being invisible to a top-level-declaration scan.

Gates: fmt, ratchet (`just refresh` — pure progress), abi-ledger, build.
scrverify txt/attrs/vals IDENTICAL; 366 functional ui/, 274 oldtest,
771 unit, 593 cargo.

+++ crates/nvim/src/nvim/bufwrite/backup.rs
@@ -18,6 +18,7 @@ use core::ffi::{c_char, c_int, c_uint};
 
 #[allow(unused_imports)]
 use super::*;
+use crate::src::nvim::highlight_group::HLF_E;
 
 /// `st_mode & S_IFMT` for a regular file.
 const S_IFREG: uint64_t = 0o100000;
@@ -68,7 +69,7 @@ unsafe fn check_mtime(buf: *mut buf_T, file_info: *mut FileInfo) -> bool {
         // Not emsg(): that would flush the buffers.
         msg(
             translate(c"WARNING: The file has been changed since reading it!!!").as_ptr(),
-            HLF_E as c_int,
+            HLF_E,
         );
         if ask_yesno(translate(c"Do you really want to write to it").as_ptr()) == 'n' as c_int {
             return false;
